+++ ios/Classes/EMGroup+Flutter.m
@@ -22,7 +22,6 @@
     ret[@"adminList"] = self.adminList;
     ret[@"blockList"] = self.blacklist;
     ret[@"muteList"] = self.muteList;
-    ret[@"sharedFileList"] = self.sharedFileList;
     ret[@"noticeEnable"] = @(self.isPushNotificationEnabled);
     ret[@"messageBlocked"] = @(self.isBlocked);
     ret[@"isAllMemberMuted"] = @(self.isMuteAllMembers);