+++ lib/src/models/em_conversation.dart
@@ -5,12 +5,6 @@ import '../tools/em_extension.dart';
 import '../../im_flutter_sdk.dart';
 import '../chat_method_keys.dart';
 
-enum EMConversationType {
-  Chat, // 单聊消息
-  GroupChat, // 群聊消息
-  ChatRoom, // 聊天室消息
-}
-
 enum EMMessageSearchDirection { Up, Down }
 
 class EMConversation {
@@ -48,10 +42,10 @@ class EMConversation {
     return data;
   }
 
-  String id = '';
-  EMConversationType? type;
+  late String id;
+  late EMConversationType type;
 
-  int? _unreadCount;
+  int _unreadCount = 0;
   Map<String, String>? _ext;
   String? _name;
   EMMessage? _latestMessage;
@@ -95,7 +89,7 @@ extension EMConversationExtension on EMConversation {
   static const MethodChannel _emConversationChannel =
       const MethodChannel('com.chat.im/chat_conversation', JSONMethodCodec());
 
-  int? get unreadCount {
+  int get unreadCount {
     return _unreadCount;
   }