+++ lib/src/tools/em_extension.dart
@@ -1,4 +1,5 @@
 // 思考： 是否要把所有格式转换的部分都放到这个extension中？
+import '../models/em_group_shared_file.dart';
 
 extension MapExtension on Map {
   bool boolValue(String key) {
@@ -12,6 +13,58 @@ extension MapExtension on Map {
       return this[key];
     } else {
       return false;
+    }
+  }
+
+  int? intValue(String key) {
+    if (this.containsKey(key)) {
+      return this[key];
+    } else {
+      return null;
+    }
+  }
+
+  String? stringValue(String key) {
+    if (this.containsKey(key)) {
+      return this[key];
+    } else {
+      return null;
+    }
+  }
+
+  List<T>? listValue<T>(String key) {
+    if (this.containsKey(key)) {
+      List obj = this[key];
+      if (T is String) {
+        List<String> strList = [];
+        for (var item in obj) {
+          strList.add(item);
+        }
+        return strList as List<T>;
+      } else if (T is EMGroupSharedFile) {
+        List<EMGroupSharedFile> fileList = [];
+        for (var item in obj) {
+          var file = EMGroupSharedFile.fromJson(item);
+          fileList.add(file);
+        }
+        return fileList as List<T>;
+      }
+    } else {
+      return null;
+    }
+  }
+}
+
+extension MapWithoutNull on Map {
+  setValueWithOutNull<T>(String key, T? value,
+      [Object Function(T object)? callback]) {
+    if (value != null) {
+      if (callback != null) {
+        Object v = callback(value);
+        this[key] = v;
+      } else {
+        this[key] = value;
+      }
     }
   }
 }